Move renderToHTMLWithComponents into a new file
